+++ CHANGELOG.md
@@ -12,3 +12,31 @@
 
 `[BUGFIX]`
 
+# 0.5.1
+
+> This is the first release of OpenFunction CLI
+
+`[CHANGE]`
+
+`[FEATURE]`
+
+`[ENHANCEMENT]`
+
+`[BUGFIX]`
+- Fix the issue that ofn install cannot install the latest version of OpenFunction #19
+
+# 0.5.2
+
+> This is the first release of OpenFunction CLI
+
+`[CHANGE]`
+
+`[FEATURE]`
+
+`[ENHANCEMENT]`
+- Change the default version of OpenFunction to the latest stable version
+- Add the `--force` option to force the operation
+- Use the spinner instead of the original process display
+- Adjust the function hierarchy to make some functions more generic
+
+`[BUGFIX]`